fix merge and move evm_perpetual to gateway module
