+++ hummingbot/connector/gateway/amm/gateway_evm_amm_lp.py
@@ -10,7 +10,7 @@ from typing import TYPE_CHECKING, Any, Dict, List, Optional, Set, Type, Union, c
 from async_timeout import timeout
 
 from hummingbot.connector.connector_base import ConnectorBase
-from hummingbot.connector.gateway_in_flight_lp_order import GatewayInFlightLPOrder
+from hummingbot.connector.gateway.amm.gateway_in_flight_lp_order import GatewayInFlightLPOrder
 from hummingbot.core.data_type.cancellation_result import CancellationResult
 from hummingbot.core.data_type.in_flight_order import OrderState
 from hummingbot.core.data_type.limit_order import LimitOrder
