+++ test/hummingbot/connector/gateway/amm/debug_gateway_evm_amm_lp.py
@@ -18,7 +18,7 @@ from typing import Generator, List, Optional
 from bin import path_util  # noqa: F401
 from hummingbot.client.config.config_helpers import read_system_configs_from_yml
 from hummingbot.connector.gateway.amm.gateway_evm_amm_lp import GatewayEVMAMMLP
-from hummingbot.connector.gateway_in_flight_lp_order import GatewayInFlightLPOrder
+from hummingbot.connector.gateway.amm.gateway_in_flight_lp_order import GatewayInFlightLPOrder
 from hummingbot.core.clock import Clock, ClockMode
 from hummingbot.core.event.event_logger import EventLogger
 from hummingbot.core.event.events import (
