+++ test/hummingbot/strategy/uniswap_v3_lp/test_uniswap_v3_lp.py
@@ -10,7 +10,7 @@ from aiounittest import async_test
 from hummingbot.client.config.client_config_map import ClientConfigMap
 from hummingbot.client.config.config_helpers import ClientConfigAdapter
 from hummingbot.connector.connector_base import ConnectorBase
-from hummingbot.connector.gateway_in_flight_lp_order import GatewayInFlightLPOrder
+from hummingbot.connector.gateway.amm.gateway_in_flight_lp_order import GatewayInFlightLPOrder
 from hummingbot.core.clock import Clock, ClockMode
 from hummingbot.core.data_type.in_flight_order import OrderState
 from hummingbot.core.data_type.trade_fee import TokenAmount, TradeFeeSchema
